requirements.txt
numpy
scipy
opencv-python
Pillow